+++ src/index.css
@@ -7,10 +7,3 @@ body {
     Arial, sans-serif;
   margin: 0;
 }
-
-#webview {
-  display: inline-flex;
-  background: #eee;
-  height: calc(100vh - 47px); /* app height - app bar height */
-  width: 100vw;
-}
