+++ src/index.html
@@ -14,9 +14,6 @@
     <title>NPLD Player</title>
   </head>
   <body>
-    <app-bar></app-bar>
-    <main>
-      <webview id="webview" src="https://www.github.com/"></webview>
-    </main>
+    <npld-player></npld-player>
   </body>
 </html>
