+++ src/renderer.ts
@@ -9,7 +9,7 @@ import '@shoelace-style/shoelace/dist/themes/light.css';
 import '@shoelace-style/shoelace/dist/shoelace.js';
 import { setBasePath } from '@shoelace-style/shoelace/dist/utilities/base-path.js';
 import './index.css';
-import './app-bar';
+import './npld-player';
 
 // Set the base path to the folder you copied Shoelace's assets to
 setBasePath('public/shoelace');
